backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.0.3
email-validator==2.1.0

# Database
prisma==0.11.0
asyncpg==0.29.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Image processing and AI/ML
numpy==1.26.4
Pillow==10.1.0
opencv-python==4.8.1.78

# HTTP client
httpx==0.25.2
aiofiles==23.2.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Environment and configuration
python-dotenv==1.0.0

# Monitoring and logging
structlog==23.2.0